feat(dashboards): add `dashboards` application
going forward this will replace the original (corrupted) `superset`
application. in time we will delete `superset` altogether, however we
want to ensure we can duplicate the existing work first. additionally
it's not possible to rename a resources so we can't just rename
`superset` to `superset-bak` or anything of the sorts.